Revert "Merge remote-tracking branch 'remotes/origin/develop' into fix/gcc_6.x"
This reverts commit 632c8fb262c279bdb231b13e730907d8741a6629, reversing
changes made to b27e9b2570a2351d193594deb87d88898bae10cd.

+++ user/tests/libraries/unit-test/unit-test.h
@@ -451,7 +451,7 @@ class Test
   // allows for both ram/progmem based names
   class TestString : public Printable {
   public:
-    const uintptr_t data;
+    const uint32_t data;
     TestString(const char *_data);
     void read(void *destination, uint16_t offset, uint8_t length) const;
     uint16_t length() const;
